add more named exports
