+++ src/createElement.ts
@@ -7,134 +7,19 @@
  * @version 1.7.0
  * @exports createElement function for DOM manipulation
  */
+// Commented this out to make tsc happy
+// eslint-disable-next-line
+// <reference types="../jsx" />
 
-import DeStagnate from "."
+import {
+    ChildrenArrayType,
+    ChildrenType,
+    EventFunc,
+    bindChildren as _bindChildren,
+    bindProps as _bindProps,
+    unpackChildren as _unpackChildren
+} from "./_createElementTools"
 import type {Ref} from "./createRef"
-
-/* eslint-disable one-var, @typescript-eslint/no-explicit-any */
-
-export type ChildrenFlatArrayType = (HTMLElement
-    | Element
-    | number
-    | string
-    | DeStagnate<any, any>
-)[]
-
-export type ChildrenArrayType = ChildrenFlatArrayType
-    | ChildrenArrayType[]
-
-/**
- * All types the children parameter can be
- */
-export type ChildrenType = HTMLElement
-    | string 
-    | number 
-    | ChildrenArrayType
-    | Element
-    | DeStagnate<any, any>
-
-export type EventFunc = (e: Event)=> void
-
-/**
- * Binds children to element
- * @package
- * @param element - element to bind
- * @param props - props to bind with
- * @param ns - if namespace element
- * @returns void
- */
-export const _bindProps = (
-    element: Element,
-    props?: {[key: string]: string | number | Element | Ref | EventFunc} | null,
-    ns = false,
-): void => {
-    if (props) {
-        for (const [key, val] of Object.entries(props)) {
-            if (typeof(val) === "string" || typeof(val) === "number") {
-                if (key === "innerHTML") {
-                    element.innerHTML = val.toString()
-                } else if (ns) {
-                    element.setAttributeNS(null, key, val.toString())
-                } else {
-                    element.setAttribute(key, val.toString())
-                }
-            } else if (key.slice(0, 2) === "on") { // Works such as onClick, onAnimationEnd, etc.
-                if (typeof(val) === "function") {
-                    element.addEventListener(
-                        key.slice(2)
-                            .toLowerCase() as keyof HTMLElementEventMap,
-                        val
-                    )
-                }
-            } else if (
-                key === "ref" &&
-                typeof(val) === "object" &&
-                "current" in val
-            ) {
-                (val as Ref<Element>).current = element
-            } else {
-                console.warn(`WARN: Invalid prop type "${typeof(val)}" for key "${key}". Skipping prop.`)
-            }
-        }
-    }
-}
-
-export const _unpackChildren = (
-    children: ChildrenArrayType,
-): ChildrenFlatArrayType => {
-    const newChildren = []
-    
-    for (const child of children) {
-        if (typeof(child) === "object" && child instanceof Array) {
-            newChildren.push(..._unpackChildren(child))
-        } else {
-            newChildren.push(child)
-        }
-    }
-
-    return newChildren as ChildrenFlatArrayType
-}
-
-/**
- * Binds children to element
- * @package
- * @param element - element to bind
- * @param children - children to bind with
- * @returns void
- */
-export const _bindChildren = (
-    element: Element,
-    children?: ChildrenType,
-): void => {
-    if (children !== null && children !== undefined) {
-        if (children instanceof Array) {
-            for (const child of children) {
-                _bindChildren(element, child)
-            }
-        } else if (
-            typeof(children) === "string" ||
-            typeof(children) === "number"
-        ) {
-            (element as HTMLElement).innerText = children.toString()
-        } else if (children instanceof DeStagnate) {
-            if (!children.didMount && element instanceof window.HTMLElement) {
-                children.mount(element)
-                
-                return
-            } else if (!(element instanceof window.HTMLElement)) {
-                throw new Error("Cannot use non-HTMLElement as component parent")
-            }
-
-            if (children.parent !== element) {
-                children.parent = element
-            }
-
-            children.forceUpdate()
-        } else {
-            element.appendChild(children)
-        }
-    }
-}
 
 /**
  * Creates a child element to DynamComponent
@@ -144,7 +29,7 @@ export const _bindChildren = (
  * @param childrenArgs - rest parameter of children
  * @returns element
  */
-const createElement = <T extends keyof HTMLElementTagNameMap>(
+export const createElement = <T extends keyof HTMLElementTagNameMap>(
     tagName: T,
     props?: {[key: string]: string | number | Element | Ref | EventFunc} | null,
     children?: ChildrenType,
